+++ now/constants.py
@@ -4,9 +4,9 @@ from docarray.typing import Image, Text, Video
 
 from now.utils import BetterEnum
 
-DOCKER_BFF_PLAYGROUND_TAG = '0.0.144-fix-bucket-uri-json-field-10'
-NOW_PREPROCESSOR_VERSION = '0.0.120-fix-bucket-uri-json-field-10'
-NOW_ELASTIC_INDEXER_VERSION = '0.0.142-fix-bucket-uri-json-field-10'
+DOCKER_BFF_PLAYGROUND_TAG = '0.0.144-fix-bucket-uri-json-field-11'
+NOW_PREPROCESSOR_VERSION = '0.0.120-fix-bucket-uri-json-field-11'
+NOW_ELASTIC_INDEXER_VERSION = '0.0.142-fix-bucket-uri-json-field-11'
 NOW_AUTOCOMPLETE_VERSION = '0.0.8-feat-one-search-dynamic-flow-66'
