+++ now/executor/preprocessor/executor.py
@@ -164,7 +164,19 @@ class NOWPreprocessor(Executor):
             parameters['user_input'] if 'user_input' in parameters else None,
         )
         self._set_user_input(parameters=parameters)
+        self.patch_docarray(docs)
         return self._preprocess_maybe_cloud_download(docs=docs)
+
+    def patch_docarray(self, docs: DocumentArray):
+        """This function modifies the documents to make sure they support features that we are missing at the moment"""
+        # 1. each attribute of the multi-modal document should know it's name
+        user_input = self.user_input.field_names_to_dataclass_fields
+        dataclass_fields_to_field_names = {v: k for k, v in user_input.items()}
+
+        for doc in docs:
+            for dataclass_field in doc._metadata['multi_modal_schema']:
+                field_name = dataclass_fields_to_field_names.get(dataclass_field, None)
+                getattr(doc, dataclass_field).tags['field_name'] = field_name
 
     @secure_request(on='/temp_link_cloud_bucket', level=SecurityLevel.USER)
     def temporary_link_from_cloud_bucket(
